@@ -4,49 +4,80 @@
   "description": "Converts an mhtml document to a single html document",
   "author": "Mayank Sindwani",
   "license": "MIT",
-  "homepage": "https://github.com/msindwan/mhtml2html#readme",
+  "type": "module",
+  "homepage": "https://github.com/puntowav/mhtml2html#readme",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/msindwan/mhtml2html.git"
+    "url": "git+https://github.com/puntowav/mhtml2html.git"
   },
   "files": [
     "dist/",
+    "src/",
     "index.js",
     "LICENSE.txt",
     "README.md"
   ],
-  "main": "dist/mhtml2html.js",
+  "main": "dist/mhtml2html.cjs",
+  "module": "dist/mhtml2html.esm.js",
+  "browser": "dist/mhtml2html.iife.js",
+  "exports": {
+    ".": {
+      "import": "./dist/mhtml2html.esm.js",
+      "require": "./dist/mhtml2html.cjs",
+      "browser": "./dist/mhtml2html.iife.js"
+    }
+  },
   "bin": {
-    "mhtml2html": "./index.js"
+    "mhtml2html": "./index.js",
+    "mhtml2html-compare": "./screenshot-compare.js"
   },
   "scripts": {
-    "lint": "eslint src",
-    "prebuild": "yarn lint",
-    "build": "webpack --mode=production",
-    "pretest": "yarn build",
-    "test": "yarn mocha tests/ && karma start --single-run"
+    "lint": "eslint src tests",
+    "lint:fix": "eslint src tests --fix",
+    "format": "prettier --write \"src/**/*.js\" \"tests/**/*.js\" \"*.js\"",
+    "format:check": "prettier --check \"src/**/*.js\" \"tests/**/*.js\" \"*.js\"",
+    "prebuild": "npm run lint",
+    "build": "npm run build:esm && npm run build:cjs && npm run build:iife",
+    "build:esm": "esbuild src/mhtml2html.js --bundle --format=esm --platform=node --outfile=dist/mhtml2html.esm.js --external:quoted-printable --external:base-64",
+    "build:cjs": "esbuild src/mhtml2html.js --bundle --format=cjs --platform=node --outfile=dist/mhtml2html.cjs --external:quoted-printable --external:base-64",
+    "build:iife": "esbuild src/mhtml2html.js --bundle --format=iife --global-name=mhtml2html --platform=browser --outfile=dist/mhtml2html.iife.js",
+    "prepare": "npm run build:only",
+    "build:only": "npm run build:esm && npm run build:cjs && npm run build:iife",
+    "pretest": "npm run build",
+    "test": "mocha tests/mhtml2html.test.js tests/chromium-exports.test.js tests/unit-helpers.test.js --timeout 30000 && karma start karma.conf.cjs --single-run",
+    "test:node": "mocha tests/mhtml2html.test.js tests/chromium-exports.test.js tests/unit-helpers.test.js --timeout 30000",
+    "test:browser": "karma start karma.conf.cjs --single-run",
+    "chromium:generate": "node tests/fixtures/generate-mhtml.js",
+    "chromium:compare": "node tests/fixtures/compare-visual.js",
+    "chromium:report": "open tests/fixtures/comparison-report.html",
+    "compare": "node screenshot-compare.js",
+    "compare:url": "node screenshot-compare.js --url",
+    "compare:urls": "node screenshot-compare.js --urls"
   },
   "engines": {
-    "node": ">=10"
+    "node": ">=18"
   },
   "dependencies": {
-    "base-64": "^0.1.0",
-    "jsdom": "^15.1.1",
+    "base-64": "^1.0.0",
+    "esbuild": "^0.20.0",
+    "jsdom": "^24.0.0",
     "quoted-printable": "^1.0.1",
-    "yargs": "^15.1.0"
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.0",
-    "babel-loader": "^8.0.6",
-    "chai": "^4.2.0",
-    "eslint": "^6.3.0",
-    "karma": "^4.3.0",
+    "chai": "^4.4.1",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "karma": "^6.4.3",
     "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^3.1.0",
+    "karma-chrome-launcher": "^3.2.0",
     "karma-cli": "^2.0.0",
-    "karma-mocha": "^1.3.0",
-    "mocha": "^6.2.0",
-    "webpack": "^4.39.3",
-    "webpack-cli": "^3.3.8"
+    "karma-mocha": "^2.0.1",
+    "mocha": "^10.3.0",
+    "pixelmatch": "^5.3.0",
+    "pngjs": "^7.0.0",
+    "prettier": "^3.2.5",
+    "puppeteer": "^22.0.0"
   }
 }
